+++ .gitignore
@@ -36,3 +36,4 @@ addon.local.gradle.kts
 addon.late.local.gradle
 addon.late.local.gradle.kts
 layout.json
+dependencies.gradle
