+++ src/main/java/twilightforest/gt_integration/GT_OrePlacer.java
@@ -1,16 +1,15 @@
-package twilightforest;
+package twilightforest.gt_integration;
 
 import net.minecraft.block.Block;
 import net.minecraft.init.Blocks;
 import net.minecraft.world.World;
 
 import gregtech.common.blocks.GT_TileEntity_Ores;
+import twilightforest.TwilightForestMod;
 
-public class GT_Integration_Utils {
+public class GT_OrePlacer {
 
-    public static boolean isIntegrationEnabled = false;
-
-    public static int getMappedGTMetaForOre(Block mOre) {
+    public int getMappedGTMetaForOre(Block mOre) {
 
         if (mOre == Blocks.coal_ore) {
             return TwilightForestMod.GT_coalOreMeta;
@@ -42,27 +41,14 @@ public class GT_Integration_Utils {
         return -1;
     }
 
-    public static boolean placeGTOre(World aWorld, int aX, int aY, int aZ, Block mcOreBlock) {
-
-        if (!isIntegrationEnabled) {
-            return false;
-        }
+    public boolean placeGTOre(World aWorld, int aX, int aY, int aZ, Block mcOreBlock) {
 
         int mappedMeta = getMappedGTMetaForOre(mcOreBlock);
-        // Something wrong
         if (mappedMeta <= 0) {
             return false;
         }
-
         return GT_TileEntity_Ores.setOreBlock(aWorld, aX, aY, aZ, mappedMeta, TwilightForestMod.GT_useSmallOres, true);
 
     }
 
-    public static void init() {
-        isIntegrationEnabled = true;
-    }
-
-    public static boolean isInit() {
-        return isIntegrationEnabled;
-    }
 }
